dsl_util: extract string formatting and parsing primitives to core
The extracted functions don't depend on the pest parsing library. They are
conceptually parts of the dsl_util module, so they are reexported from there.
