+++ core/src/lib.rs
@@ -39,6 +39,7 @@ pub mod merge;
 pub mod object_id;
 pub mod repo_path;
 pub mod str_util;
+pub mod symbol_util;
 
 #[cfg(test)]
 mod tests {
